+++ packages/headless/src/test/mock-citation.ts
@@ -1,0 +1,21 @@
+import {GeneratedAnswerCitation} from '../controllers';
+
+/**
+ * For internal use only.
+ *
+ * Returns a `GeneratedAnswerCitation` for testing purposes.
+ * @param config  - A partial `GeneratedAnswerCitation` from which to build the target `GeneratedAnswerCitation`.
+ * @returns The new `GeneratedAnswerCitation`.
+ */
+export function buildMockCitation(
+  config: Partial<GeneratedAnswerCitation> = {}
+): GeneratedAnswerCitation {
+  return {
+    id: '',
+    title: '',
+    uri: '',
+    permanentid: '',
+    clickUri: '',
+    ...config,
+  };
+}